Fix comment count, comment form rendering order, & duplicate comment form under berries
